requirements.txt
gunicorn
setproctitle
Flask
flask_sqlalchemy
flask_cors
--allow-external mysql-connector-python
mysql-connector-python
python_jwt
requests
requests-oauthlib
eventlet>=0.17.4

flask_restful
pytz
aniso8601

flask_socketio
flask_redis

apns-clerk

pillow

python_dateutil

datadog

Flask-Migrate
Flask-Admin
Flask-Login